modified:   draft/392-string-trim.cc 	new file:   draft/784-wonderful-sub-array.cc 	modified:   draft/834-legend-of-the-galactic-heros.cc 	deleted:    draft/862-card-game.cc 	modified:   draft/863-stern-brocot-tree.cc 	modified:   draft/d.cc 	modified:   draft/t.cc 	renamed:    draft/211-points.cc -> solution/211-points.cc 	new file:   solution/356-poor-trader-advisor.cc 	new file:   solution/862-card-game.cc

+++ solution/211-points.cc
@@ -3,10 +3,7 @@
 //
 //#include <stdio.h>
 #include <iostream>
-//#include <array>
-//#include <string>
 #include <vector>
-//#include <unoredered_map>
 #include <algorithm>
 #include <math.h>
 #define endl "\n"
@@ -14,7 +11,7 @@ using namespace std;
 
 struct point
 {
-    int x, y;
+    double x, y; // double matters!!!!
 }p, p0;
 
 int t, pos;
